# os/bluestore: Refactor Bluestore Caches