@tchaikov 2019-06-20
http://pulpito.ceph.com/kchai-2019-06-20_05:33:07-rados-wip-kefu-testing-2019-06-20-1103-distro-basic-smithi/

the valgrind failures are caused by #27834