@markhpc 2019-06-24
New changes:

nits cleanup
moved dumped_onodes into the onode cache
external tracking of cache size using an atomic to avoid lock for stats
removed trim from discard and only trim when we lock (address Igor's previous comments)

Performance didn't appear to change significantly:

wip-bs-ind-cache:

before: 39564 IOPS
after: 39008 IOPS

wip-bs-ind-cache + #27705:

before: 43219 IOPS
after: 43286 IOPS

Inverse wallclock profiles attached for the base and with #22705 cases.  There's still some lock contention on the cache in the tp_osd_tp threads (#22705 helps by avoiding double caching) but now we're hitting pglock as well.  Bigger fish to fry may be memcpy in BlueRocksWritableFile::Append.
[wip-bs-ind-cache-update.inverse.txt](https://github.com/ceph/ceph/files/3322840/wip-bs-ind-cache-update.inverse.txt)
[wip-bs-ind-cache-update-27705.inverse.txt](https://github.com/ceph/ceph/files/3322841/wip-bs-ind-cache-update-27705.inverse.txt)