@markhpc 2019-07-02
Looking at:

http://pulpito.ceph.com/sage-2019-06-26_20:55:20-rados-wip-sage-testing-2019-06-26-0958-distro-basic-smithi/

I see a ton of failures due to ceph-objecstore-tool being missing.  Possibly several with scrubbing terminating with not all pgs active/clean.