@markhpc 2019-07-23
New run results:

http://pulpito.ceph.com/nhm-2019-07-22_23:44:17-rados-wip-mark-testing-2019-07-22-0826-distro-basic-smithi/