@markhpc 2019-07-26
http://pulpito.ceph.com/nhm-2019-07-25_12:34:31-rados-wip-mark-testing-2019-07-24-0720-distro-basic-smithi/
http://pulpito.ceph.com/nhm-2019-07-25_00:46:08-rados-wip-mark-testing-2019-07-24-0720-distro-basic-smithi/